fixed chakra stuff
